Correction problème de distance max

+++ NeuralNetwork/Config/SensorConfig.cs
@@ -9,6 +9,6 @@ namespace NeuralNetwork.Config
     public static class SensorConfig
     {
         public static int SIDE_LENGHT = 100;
-        public static int SENSOR_LENGHT { get; } = ((int)Math.Sqrt(2)) * SIDE_LENGHT; // Lenght of the sensor ==> lenght = 71 = sqrt(2) * SIDE_LENGHT
+        public static double SENSOR_LENGHT { get; } = ((double)Math.Sqrt(2)) * SIDE_LENGHT; // Lenght of the sensor ==> lenght = sqrt(2) * SIDE_LENGHT
     }
 }
